API/requirements.txt

Django==4.1.7
gunicorn==20.1.0
djangorestframework==3.14.0
python-decouple==3.8
django-cors-headers==4.2.0
pandas==2.2.3
joblib==1.4.2
psycopg2-binary==2.9.9
openpyxl==3.1.2
cryptography==41.0.2
pyjwt==2.8.0
prophet==1.1.6